refactor: split stores
