+++ .env
@@ -1,5 +1,5 @@
-VITE_AZURE_OPEN_AI_API_KEY= # Replace with your API key
-VITE_AZURE_OPEN_AI_API_VERSION= # Replace with the API version
-VITE_AZURE_OPEN_AI_DEPLOYMENT= # Replace with the deployment name of your OpenAI model
-VITE_AZURE_OPEN_AI_ENDPOINT= # Replace with the Azure OpenAI endpoints
-VITE_AZURE_OPEN_AI_MODEL= # Replace with the model name
+VITE_AZURE_OPEN_AI_API_KEY=832d4d21a5ee4e6ab33f03f1ac47b0d8 # Replace with your API key
+VITE_AZURE_OPEN_AI_API_VERSION=2023-03-15-preview # Replace with the API version
+VITE_AZURE_OPEN_AI_DEPLOYMENT=aida-apps-gpt-4o # Replace with the deployment name of your OpenAI model
+VITE_AZURE_OPEN_AI_ENDPOINT=https://services-aida-apps.openai.azure.com # Replace with the Azure OpenAI endpoints
+VITE_AZURE_OPEN_AI_MODEL=gpt-4o # Replace with the model name
