+++ src/hooks/use-refine-detail-backlog.tsx
@@ -5,8 +5,9 @@ import { useToast } from "@/components/toaster/hook/use-toast";
 import { ToastVariant } from "@/components/ui/toast/toast.const";
 import useGetDetailRefinedFeature from "@/features/backlog-feature/hooks/use-detail-refined-backlog";
 import { transformDetailsToString } from "@/helpers/helpers";
-import { useAppStore } from "@/hooks/use-app-store";
 import { parseAidaRefinedDetailBacklogResponse } from "@/lib/aida";
+import { useAppStore } from "@/stores/use-app-store";
+import { useBacklogFeatureStore } from "@/stores/use-backlog-feature-store";
 import { RefinedBacklogDetails } from "@/types/common";
 import { sendRefinedDetailBacklog } from "@/utils/utils-aida-service";
 
@@ -14,7 +15,8 @@ const useRefineDetailBacklog = () => {
   const { t } = useTranslation();
   const { toast } = useToast();
 
-  const { feature, updateRefinedBacklog, setIsLoadingDetail } = useAppStore();
+  const { feature } = useAppStore();
+  const { updateRefinedBacklog, setIsLoadingDetail } = useBacklogFeatureStore();
 
   const { detailRefinedBacklog } = useGetDetailRefinedFeature();
