+++ src/hooks/use-refine-feature-name.tsx
@@ -4,21 +4,17 @@ import { useTranslation } from "react-i18next";
 import { useToast } from "@/components/toaster/hook/use-toast";
 import { ToastVariant } from "@/components/ui/toast/toast.const";
 import { mapRefinedFeatureResponse } from "@/helpers/helpers";
-import { useAppStore } from "@/hooks/use-app-store";
 import { parseAidaRefinedFeatureResponse } from "@/lib/aida";
+import { useAppStore } from "@/stores/use-app-store";
+import { useRefineFeatureStore } from "@/stores/use-refine-feature-store";
 import { sendRefinedFeatureName } from "@/utils/utils-aida-service";
 
 const useRefineFeatureName = () => {
   const { t } = useTranslation();
   const { toast } = useToast();
 
-  const {
-    feature,
-    updateFeature,
-    refinedFeature,
-    setRefinedFeature,
-    setIsLoading,
-  } = useAppStore();
+  const { feature, updateFeature, setIsLoading } = useAppStore();
+  const { refinedFeature, setRefinedFeature } = useRefineFeatureStore();
 
   const fetchRefinedFeatureName = useCallback(async () => {
     try {
